% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{data.frame of all available countries and their ISO-3166 country codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 256 rows and 7 columns.
}
\usage{
country_codes
}
\description{
data.frame of all available countries and their ISO-3166 country codes
}
\keyword{datasets}
