% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{flags}
\alias{flags}
\title{Retrieve the link to image files for flag}
\usage{
flags(ccodes, filetype = c("png", "svg"))
}
\arguments{
\item{ccodes}{Country codes in ISO-3166 format e.g. 'au' for Australia}

\item{filetype}{one of 'png' or 'svg'.  Default: 'png'}
}
\value{
character string of filenames.  If an invalid country code is given,
or there was a problem finding the file, an \code{NA_character_} will be
returned at this location.
}
\description{
Retrieve the link to image files for flag
}
